Excel Controller in added

+++ src/main/java/org/example/controllers/CatagoryController.java
@@ -3,6 +3,7 @@ package org.example.controllers;
 import lombok.AllArgsConstructor;
 import org.example.entities.Catagory;
 import org.example.repository.CatagoryRepository;
+import org.example.repository.ProductRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,6 +15,7 @@ import java.util.List;
 public class CatagoryController {
 
     private final CatagoryRepository catagoryRepository;
+    private final ProductRepository productRepository;
 
     @GetMapping("/catagories")
     public ResponseEntity<List<Catagory>> getAllCatagories() {
@@ -40,6 +42,9 @@ public class CatagoryController {
 
     @DeleteMapping("catagories/{id}")
     public ResponseEntity<String> deleteCatagory(@PathVariable Integer id) {
+        if(productRepository.countByCatagoryId(id) > 0) {
+            return ResponseEntity.badRequest().body("Cannot delete category with existing products.");
+        }
         catagoryRepository.deleteById(id);
         return ResponseEntity.ok("Successfully Deleted");
     }
